add doubletap2wake, completely rewritten on the s2w template
This is as easy to port as s2w, if not easier since no touchscreen
coords are necessary for it to work.

Just throw it in your tree, force the touchscreen to stay online and
you are done.

Signed-off-by: Dennis Rassmann <showp1984@gmail.com>
Signed-off-by: flar2 <asegaert@gmail.com>

(cherry picked from commit f9573747bdea5a64d237db16cb6d0a239c731b0b)